Fix gulp test task
jest.runCLI() was being called with only 1 argument, there was a mis-bracketing
which lead to this. When trying to fix the bracketing here, I realised the
error handling callback also did not work correctly (done was undefined), and
the 'test' task considered itself succeeded before the tests were even run.

I've refactored the task a little to successfully "fail" when the tests fail,
which means that "gulp test" will now return a non-zero exit code when the
tests fail.